+++ prob_map_view/include/prob_map_view/functions.h
@@ -12,10 +12,10 @@ inline T pToL(T p) {
   return std::log(p / (1-p));
 }
 
-inline cv::Mat_<double> pToL(cv::Mat_<double> p, double min_log){
-  cv::Mat_<double> res;
+inline cv::Mat_<float> pToL(cv::Mat_<float> p, float min_log){
+  cv::Mat_<float> res;
   p.copyTo(res);
-  for (double &v : cv::Mat_<double>(res)) {
+  for (float &v : cv::Mat_<float>(res)) {
     v = std::log10(v); v = (v > min_log ? (-min_log+v)/(-min_log) : 0.0);
   }
   return res;
@@ -26,8 +26,8 @@ inline T lToP(T l) {
   return 1 - 1 / (1 + std::exp(l));
 }
 template<>
-inline cv::Mat_<double> lToP(cv::Mat_<double> l){
-  cv::Mat_<double> tmp;
+inline cv::Mat_<float> lToP(cv::Mat_<float> l){
+  cv::Mat_<float> tmp;
   cv::exp(l, tmp);
   return tmp / (1+tmp);
 }
